feat: add keysend option to manual send flow
this allows users to do keysend payments from the send screen
